Refactor: hook 호출 위치 수정
